# Optimise BaseExpression creation