--- a/mathics/settings.py
+++ b/mathics/settings.py
@@ -27,9 +27,6 @@ TIMEOUT = None
 
 MAX_RECURSION_DEPTH = 512
 
-# number of bits of precision for inexact calculations
-MACHINE_PRECISION = 64
-
 ADMINS = (
     ('Admin', 'mail@test.com'),
 )