Some small changes to the core to speed up `BaseExpression` creation.

Also includes some enhancements to the benchmarking script.

### Before:
```
'Expand[(a1+a2+a3+a4+a5+a6+a7)^3]'
    10 loops, avg:  147 ms, best:  136 ms, median:  145 ms per loop
%timeit Integer(5)
    1000000 loops, best of 3: 1.52 µs per loop
%timeit Expression('Sin')
     100000 loops, best of 3: 4.03 µs per loop
```

### After:
```
'Expand[(a1+a2+a3+a4+a5+a6+a7)^3]'
      100 loops, avg: 98.3 ms, best: 93.1 ms, median: 95.5 ms per loop
%timeit Integer(5)
    1000000 loops, best of 3: 1.15 µs per loop
%timeit Expression('Sin')
     100000 loops, best of 3: 3.2  µs per loop
```

### After `lru_cache`:  **(deferred)**
```
'Expand[(a1+a2+a3+a4+a5+a6+a7)^3]'
    100 loops, avg: 87.8 ms, best: 82.4 ms, median: 86.2 ms per loop
%timeit Integer(5)
    10000000 loops, best of 3: 158 ns per loop
%timeit Expression('Sin')
    1000000 loops, best of 3: 1.99 µs per loop
```

The introduction of LRU cache caused some issues when copying and modifying expressions so I've removed them from this PR.